Add support for simple web tables
This is a very first experimental try to add the support for web tables.

Only NavigateDown is changed for now as things are going to get slightly more interesting here.
The currently used algorithm is quite naive, it simply takes the parent of its parent and considers it to be the stack of rows, which it inspects one by one. No recursion, no sophisticated structures.

Although it works on simple, regular tables, it was really more about me trying to find out, how much information can I get about the HTML structure of the viewed page (for example whether I'm in table in first place, whether a certain node is cell, row, etc.).

The results were not quite satisfying, i was not able to get a single thing about the type of the focused element, but this is probably my fault as JSR can somehow clearly distinguish it.

Another problem is scrolling, accessibility nodes of a window for some reason do not contain nodes out of the screen, so it need to be scrolled first to get on them.
But How do I find out, that I'm on the end of the screen and not simply on the bottom of the table?

IDK

At least for now, this needs further investigation from my side. As I think of it, wouldn't it be kind of smarter to deal with scrolling first and resolve tables after?
Hmm, thinking really seems like useful activity.